move QueryContext and QueryFileInterpreter out of queryfile
